+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.5.1
 
 - 🦺 Enabling Rich Presence on Discord after a game has already started would not show the game as being played. It now correctly updates within 5 seconds.
+- 🦺 Fixed a rare case where "Run on Startup" would be checked but would not actually start. This was due to the .exe being moved after "Run on Startup" was checked.
 
 ## v1.5.0